Respect positions of pinned posts
